+++ packages/angular_devkit/core/node/experimental/jobs/index.ts
@@ -5,8 +5,4 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import * as jobs from './jobs';
-
-export {
-  jobs,
-};
+export * from './job-registry';
