+++ packages/angular_devkit/core/node/index.ts
@@ -5,7 +5,7 @@
  * Use of this source code is governed by an MIT-style license that can be
  * found in the LICENSE file at https://angular.io/license
  */
-import * as experimental from './experimental/job-registry';
+import * as experimental from './experimental/jobs/job-registry';
 import * as fs from './fs';
 export * from './cli-logger';
 export * from './host';
